+++ src/main/java/fr/softsf/sudokufx/dto/MenuDto.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
 public record MenuDto(
-        Long menuid,
+        Byte menuid,
         @NotNull
         @Min(1)
         @Max(3)
